\name{bem}
\docType{data}
\alias{bem}
\title{Binary event matrix of mutational status}
\description{
Contains 292 cell lines (rows) and 71 oncogenes (columns). The cell lines are the overlapping ones between CCLE and GDSC. The mutational status is from Garnett et al. 2012 and obtained with capillary sequencing. Row names are COSMIC identifiers and column names are HUGO gene names. 
}
\usage{data(bem)}
\format{A matrix containing 292 cell lines and 71 oncogenes}
\references{
Garnett, M.J. et al. Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature 483, 570-575 (2012).

Barretina, J. et al. The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity. Nature 483, 603-607 (2012).
}
\keyword{datasets}