\name{ccle_ActArea}
\docType{data}
\alias{ccle_ActArea}
\title{CCLE drug response data measured in area above the fitted dose response curve}
\description{
Contains the drug response from CCLE for 292 cell lines (rows) and 15 drugs (columns), which are the overlapping cell lines and drugs with GDSC. Row names containing the COSMIC cell identifiers and column names are the drug names. 
}
\usage{data(ccle_ActArea)}
\format{A matrix containing 292 cell lines and 15 drugs}
\references{
Garnett, M.J. et al. Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature 483, 570-575 (2012).

Barretina, J. et al. The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity. Nature 483, 603-607 (2012).
}
\keyword{datasets}