\name{plotComparison}
\alias{plotComparison}
\title{Compares two ANOVA outputs}
\description{
Comparison plot of two objects from \link{DRANOVA_OUT-class}, which both can be generated with \link{runANOVA}. 
}
\usage{
plotComparison(X, Y, FDR=0.2, main="", xlab="", ylab="", withLabel=TRUE, labelSig=FDR, ...)
}
\arguments{
  \item{X}{
    Object of \link{DRANOVA_MUT_OUT-class} or \link{DRANOVA_OUT-class}, which is generated by \link{runANOVA} and \link{runTissueMultiComp}, respectively. It will be compared to \code{"Y"}.}

  \item{Y}{
    Object of \link{DRANOVA_OUT-class} or \link{DRANOVA_OUT-class}, which will be compared to \code{"X"}. }
  
  \item{FDR}{False discovery rate (FDR) threshold, which is equivalent to Benjamini & Hochberg correction. Default is 0.2 (=20\% FDR).}
   
  \item{main}{Title of the comparison plot.}
  
  \item{xlab}{x-label of volcano plot.}
  
  \item{ylab}{y-label of volcano  plot.}
  
  \item{withLabel}{Boolean value if drug-to-oncogene associations should be plotted. Default is \code{"TRUE"}.}
  
  \item{labelSig}{To which FDR level labels should be plotted, e.g. 30\% FDR level = 0.3. Requires parameter \code{"withLabel"} set to be \code{"TRUE"}. Default of \code{"labelSig"} is equals setting of parameter \code{FDR}.}
  
  \item{\dots}{ ...}

}
\details{
Consistency of drug sensitivity prediction markers between the two ANOVA analysis, using \link{runANOVA}. Each point is a tested association and its position on the two axis corresponds to its 'signed log q-values', for the two datasets respectively. The sign is determined by the effect of the marker (neg = increased sensitivity and pos = increased resistance) and the modulus is the log q-value of the corresponding ANOVA test, after correcting for multiple hypothesis testing with Benjamini & Hochberg. Associations which are not significant in either dataset are displayed grey, in at least on dataset being significant are colored red or green (depending on associated with resitance or sensitity), and in case the association is identified in both screenings than furthermore it is circled black. 
}
\seealso{
  \code{\link{DRANOVA_OUT-class}},
  \code{\link{DRANOVA_MUT_OUT-class}},
  \code{\link{runANOVA}},
  \code{\link{runTissueMultiComp}},
  \code{\link{plotVolcano}}
}
\examples{
# load data
data(ccle_ActArea)  # Drug response in ActiveArea from CCLE
data(gdsc_ActArea)  # Drug response in 1-AUC from GDSC
data(bem)           # Binary event matrix (71 oncogenes)
data(tissue)        # Cell line tissue-of-origin

# run anova with drug response from CCLE
out_ccle = runANOVA(new("DRANOVA", 
                        res=ccle_ActArea, 
                        bem=bem, 
                        tissue=tissue))

# run anova with drug response from GDSC
out_gdsc = runANOVA(new("DRANOVA", 
                       res=gdsc_ActArea, 
                       bem=bem, 
                       tissue=tissue))
                       
# Comparison plot between CCLE and GDSC
plotComparison(out_gdsc, out_ccle,
               main="CCLE & GDSC comparison with Active Area",
               xlab="GDSC ANOVA signed -log10(q-value)",
               ylab="CCLE ANOVA signed -log10(q-value)")
}
\keyword{plot}