\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot of ANOVA output}
\description{
Volcano plot of an \link{DRANOVA_MUT_OUT-class} or \link{DRANOVA_OUT-class} object, which can be generated with \link{runANOVA} and \link{runTissueMultiComp}, respectively. 
}
\usage{
plotVolcano(X, FDR=0.2, main="", withLabel=TRUE, labelSig=FDR, ...)
}

\arguments{
  \item{X}{Object of \link{DRANOVA_OUT-class}, which is generated by \link{runANOVA}.}
  
  \item{FDR}{False discovery rate (FDR) threshold, which is equivalent to Benjamini & Hochberg correction. Default is 0.2 (=20\% FDR).}
  
  \item{main}{Title of the volcano plot.}
  
  \item{withLabel}{Boolean value if drug-to-oncogene associations should be plotted. Default is \code{"TRUE"}.}
  
  \item{labelSig}{To which FDR level labels should be plotted, e.g. 30\% FDR level = 0.3. Requires parameter \code{"withLabel"} set to be \code{"TRUE"}. Default of \code{"labelSig"} is equals setting of parameter \code{FDR}.}
  
  \item{\dots}{...}   
}

\details{
  Volcano plot is a representation of an analysis of variance (ANOVA). Each circle represents a drug–gene interaction (circle size is proportional to the number of cell lines screened with a mutation in the corresponding gene). The x-axis shows the magnitude of the effect (signed Cohen's D): negative values (green circle) correspond to mutations associated with an increase in sensitivity, and positive values (red circle) mutations associated with decrease in sensitivity.  The y-axis represents the significance of the associations (corrected p-value = q-values) and the horizontal dashed line indicates the chosen FDR threshold for statistical significance.}

\seealso{
\code{\link{DRANOVA_MUT_OUT-class}},
\code{\link{DRANOVA_OUT-class}},
\code{\link{runANOVA}},
\code{\link{runTissueMultiComp}}
}

\examples{
# load data
data(gdsc_ActArea)  # Drug response in 1-AUC from GDSC
data(bem)           # Binary event matrix (71 oncogenes)
data(tissue)        # Cell line tissue-of-origin

# run anova with drug response from GDSC
out_gdsc = runANOVA(new("DRANOVA", 
                       res=gdsc_ActArea, 
                       bem=bem, 
                       tissue=tissue))

# Volcano plot of GDSC
plotVolcano(out_gdsc, 
            main="GDSC",
            withLabel=TRUE,
            labelSig=0.01)
} 
\keyword{plot}
