\name{tissue}
\docType{data}
\alias{tissue}
\title{tissue-of-origin of cell lines}
\description{
Contains a vector with tissue-of-origin for the 292 cell lines. The cell lines are the overlap between CCLE and GDSC (see references), and names of the vector are COSMIC identifiers. The vector contains 16 different tissue types, e.g. blood, breast, lung, skin, etc. 
}
\usage{data(tissue)}
\format{A vector containing the tissue type of the cell lines}
\references{
Garnett, M.J. et al. Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature 483, 570-575 (2012).

Barretina, J. et al. The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity. Nature 483, 603-607 (2012).
}
\keyword{datasets}